frontend/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
pyodbc>=4.0.39
sqlalchemy>=2.0.0
requests>=2.31.0
azure-identity>=1.15.0
openai>=1.3.0
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.34.0
